+++ src/main/java/de/alpharogroup/file/zip/Zip4jExtensions.java
@@ -25,6 +25,8 @@
 package de.alpharogroup.file.zip;
 
 import java.io.File;
+import java.util.Arrays;
+import java.util.List;
 
 import lombok.experimental.UtilityClass;
 import net.lingala.zip4j.core.ZipFile;
@@ -117,19 +119,37 @@ public final class Zip4jExtensions
 	}
 
 	/**
-	 * Adds a file or several files to the given zip file.
+	 * Adds the given file(s) to the given zip file.
 	 *
 	 * @param zipFile4j
 	 *            the zip file4j
 	 * @param parameters
 	 *            the parameters
 	 * @param toAdd
-	 *            the to add
+	 *            the list with the files to add in the zip file
 	 * @throws ZipException
 	 *             the zip exception
 	 */
 	public static void zipFiles(final ZipFile zipFile4j, final ZipParameters parameters,
 		final File... toAdd) throws ZipException
+	{
+		zipFiles(zipFile4j, parameters, Arrays.asList(toAdd));
+	}
+
+	/**
+	 * Adds the given file(s) to the given zip file.
+	 *
+	 * @param zipFile4j
+	 *            the zip file4j
+	 * @param parameters
+	 *            the parameters
+	 * @param toAdd
+	 *            the list with the files to add in the zip file
+	 * @throws ZipException
+	 *             the zip exception
+	 */
+	public static void zipFiles(final ZipFile zipFile4j, final ZipParameters parameters,
+		final List<File> toAdd) throws ZipException
 	{
 		for (final File element : toAdd)
 		{